crypto_scan/requirements.txt
aio-pika==9.5.5
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiormq==6.8.1
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
ccxt==4.4.71
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clickhouse-connect==0.8.15
cryptography==44.0.2
exceptiongroup==1.2.2
frozenlist==1.5.0
idna==3.10
json-to-clickhouse==0.1.18
jsonpath-ng==1.7.0
lz4==4.4.3
multidict==6.2.0
numpy==2.2.4
pamqp==3.3.0
pandas==2.2.3
pika==1.3.2
ply==3.11
propcache==0.3.0
psutil==7.0.0
pycares==4.5.0
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
requests==2.32.3
setuptools==77.0.3
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
